deployment_vertexai/requirements.txt
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
pandas==2.3.1
numpy==2.3.2
xgboost==3.0.4
scikit-learn==1.7.1
joblib==1.5.1
flask==2.3.3
gunicorn==21.2.0
